requirements.txt
Flask==2.1.2
line-bot-sdk==2.2.1
pandas==1.4.2
gspread==5.3.2
google-auth==2.6.6
APScheduler==3.0.0
gunicorn==20.0.4